Lexibomb Server v0.0.1-dev LexibombServer.Board

Types

coord ::
  LexibombServer.Board.Grid.coord |
  {non_neg_integer, String.t} |
  String.t

seed :: {integer, integer, integer}

t :: %{grid: LexibombServer.Board.Grid.t, seed: seed}

Functions

get(pid)

Specs

get(pid) :: t

new(size \\ 15)

Specs

new(pos_integer) :: t

new(size, seed)

Specs

new(pos_integer, seed) :: t

parse_coord(coord)

Specs

parse_coord(coord) ::
  {:ok, LexibombServer.Board.Grid.coord} |
  {:error, :badarg}

Parses and normalizes a given coordinate.

It will parse any of the valid LexibombServer.Board.coord forms and will normalize them to the underlying LexibombServer.Board.Grid.coord type. If indicating the column with a string, you can use either uppercase or lowercase letters, and you may also optionally separate the row and column with whitespace.

Both row and col must parse into non-negative integers for success.

Examples

iex> LexibombServer.Board.parse_coord "10B"
{:ok, {10, 2}}
iex> LexibombServer.Board.parse_coord "10 b"
{:ok, {10, 2}}
iex> LexibombServer.Board.parse_coord {10, "B"}
{:ok, {10, 2}}
iex> LexibombServer.Board.parse_coord {10, 2}
{:ok, {10, 2}}
iex> LexibombServer.Board.parse_coord {-10, 2}
{:error, :badarg}
iex> LexibombServer.Board.parse_coord {"B", 10}
{:error, :badarg}

Returns {:ok, {row, col}} on success, or {:error, :badarg} on failure.

place_bomb(pid, coord)

Specs

place_bomb(pid, coord) :: :ok | {:error, :badarg}

Places a bomb on the board square at the given coordinate.

Returns :ok on success, or {:error, :badarg} on failure.

place_bombs(pid, coords)

Specs

place_bombs(pid, [coord]) :: :ok | {:error, :badarg}

Places a bomb on each of the board squares at the given coordinates.

Returns :ok on success, or {:error, :badarg} on failure.

place_random_bombs(pid, count \\ 22)

Specs

place_random_bombs(pid, pos_integer) :: :ok

Places a bomb on count number of randomly selected board squares.

set(board, pid)

Specs

set(t, pid) :: :ok

size(pid)

Specs

size(pid) :: pos_integer

start_link()

Specs

start_link :: Agent.on_start

start_link(board)

Specs

start_link(t) :: Agent.on_start

valid_square?(pid, coord)

Specs

valid_square?(pid, coord) :: boolean

Validates if the given coordinate points to a square on the board.