Lexibomb Server v0.0.1-dev LexibombServer.Board.Grid

Acts as the glue between the high-level board interface and the low-level square states.

The grid module is responsible for establishing and manipulating the underlying data structure of a board’s coordinate system. The squares within a grid are referred to using 0-based indexes for both the rows and columns. These squares are stored within a MapSet using a {row, col} tuple as the key.

An extra border of inactive squares is added to the a board’s size in order to simplify functions that would otherwise have to perform complicated boundary checks. This border means that a 15 x 15 sized board, will utilize a 17 x 17 grid underneath.

Types

col :: non_neg_integer

coord :: {row, col}

row :: non_neg_integer

t :: %{coord => LexibombServer.Board.Square.t}

Functions

active_squares(grid)

Specs

active_squares(t) :: t

Returns a map containing just the active squares on the given grid.

board_size(grid)

Specs

board_size(t) :: pos_integer

Returns the size of the game board.

This does not include the additional rows/cols of the deactivated border.

cascading_reveal(grid, coord)

Specs

cascading_reveal(t, coord) :: t

Reveal the square at the given coordinate; if no bombs are adjacent, recursively reveal all adjacent squares.

init(size)

Specs

init(pos_integer) :: t

Initializes a new grid of squares with a deactivated border.

place_bomb(grid, coord)

Specs

place_bomb(t, coord) :: t

Places a bomb on the grid square at the given coordinate.

place_bombs(grid, coords)

Specs

place_bombs(t, [coord]) :: t

Places a bomb on each of the grid squares at the given coordinates.

place_tile(grid, coord, tile)

Specs

place_tile(t, coord, String.t) :: t

Places a tile on the grid square at the given coordinate.

This triggers a cascading reveal.

reveal(grid, coord)

Specs

reveal(t, coord) :: t

Reveals the grid square at the given coordinate.

size(grid)

Specs

size(t) :: pos_integer

Returns the size of the grid.

This does include the additional rows/cols of the deactivated border.

valid_coord?(grid, coord)

Specs

valid_coord?(t, coord) :: boolean