Lexibomb Server v0.0.1-dev LexibombServer.Board.Square

Manages the low-level state data associated with an individual board square.

A square can either be hidden or revealed, which indicates whether its state data should be available externally. Placing a tile on a square automatically sets it to the revealed state.

Types

t :: %{adjacent_bombs: non_neg_integer, bomb?: boolean, revealed?: boolean, tile: String.t}

Functions

active?(square)

Specs

active?(t) :: boolean

Returns true if the square is active.

deactivate(square)

Specs

deactivate(t) :: t

Returns a copy of square in the inactive state.

inc_adjacent_bombs(square)

Specs

inc_adjacent_bombs(t) :: t

Returns a copy of square with an incremented adjacent bomb count.

no_adjacent_bombs?(square)

Returns true if the square has no adjacent bombs.

place_bomb(square)

Specs

place_bomb(t) :: t

Returns a copy of square with a bomb placed on it.

place_tile(square, tile)

Specs

place_tile(t, String.t) :: t

Returns a copy of square in the revealed state with tile placed on it.

playable?(square)

Returns true if the square has no tile placed on it.

reveal(square)

Specs

reveal(t) :: t

Returns a copy of square in the revealed state.

revealed?(square)

Specs

revealed?(t) :: boolean

Returns true if the square is revealed.