Lexibomb Server v0.0.1-dev LexibombServer.Play

Types

cardinal_direction :: :n | :e | :s | :w

direction :: :across | :down

t :: %{start: LexibombServer.Board.coord, direction: direction, letters: [String.t], rack_leave: [String.t]}

Functions

crossword(grid, coord, direction)

Specs

crossword(LexibombServer.Board.Grid.t, LexibombServer.Board.Grid.coord, direction) :: String.t

Returns the word that intersects coord in the perpendicular direction to the given direction.

Uses ‘.’ for the one square that is missing a letter.

letters_with_coords(play)

Specs

letters_with_coords(t) ::
  {:ok, [{String.t, LexibombServer.Board.Grid.coord}]} |
  {:error, :badarg}

Returns the list of letters with each letter wrapped in a tuple alongside its coordinate.

opposite(direction)

Specs

opposite(cardinal_direction) :: cardinal_direction

Returns the cardinal direction that is opposite to the given direction.

scan_letters(grid, coord, direction)

Specs

scan_letters(LexibombServer.Board.Grid.t, LexibombServer.Board.Grid.coord, cardinal_direction | direction) :: LexibombServer.Board.Grid.coord

scan_to_anchor(grid, coord, direction)

Specs

scan_to_anchor(LexibombServer.Board.Grid.t, LexibombServer.Board.Grid.coord, cardinal_direction | direction) :: LexibombServer.Board.Grid.coord