Lexibomb Server v0.0.1-dev LexibombServer.Rack

Provides functions related to dealing with a rack of tiles.

A rack is a list of (usually 7) tiles, like ["E", "E", "L", "R", "T", "T", "S"].

Functions

letters(rack)

Specs

letters([String.t]) :: [String.t]

Returns the distinct letters in a rack.

Includes the lowercase alphabet if there are any blank tiles.

Examples

iex> LexibombServer.Rack.letters ~W(E E L R T T S)
["E", "L", "R", "T", "S"]
iex> LexibombServer.Rack.letters ~W(E E L R T T _)
["E", "L", "R", "T", "a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l",
 "m", "n", "o", "p", "q", "r", "s", "t", "u", "v", "w", "x", "y", "z"]

remove(rack, tiles)

Specs

remove([String.t], [String.t]) :: String.t

Returns a copy of rack with the given tiles removed.

If any of the given tiles are lowercase letters, a corresponding blank tile is removed from the rack.

Examples

iex> LexibombServer.Rack.remove ~W(E E L R T T S), ~W(S E T)
["E", "L", "R", "T"]
iex> LexibombServer.Rack.remove ~W(E E L R T T _), ~W(T R E a T)
["E", "L"]