Lexibomb Server v0.0.1-dev LexibombServer.WordList

Manages access to the list of valid gameplay words.

Functions

get()

Specs

get :: MapSet.t

Retrieves the default word list from the agent.

member?(word)

Specs

member?(String.t) :: boolean

Checks if the word list contains word.

prefix?(prefix)

Specs

prefix?(String.t) :: boolean

Checks if the given prefix is a valid prefix in the word list.

prefixes(word_list)

Specs

prefixes(MapSet.t) :: MapSet.t

Returns the set of all prefixes of each word in the given word_list.

Examples

iex> word_list = MapSet.new(["HELLO", "HELP", "HELPER"])
iex> LexibombServer.WordList.prefixes(word_list)
#MapSet<["", "H", "HE", "HEL", "HELL", "HELP", "HELPE"]>

start_link(word_list \\ default_list)

Specs

start_link(MapSet.t) :: Agent.on_start

Starts an agent linked to the current process to store a normalized version of word_list.

The default word list is based on Mendel Leo Cooper’s Yet Another Word List (YAWL) project.