Lexibomb Server v0.0.1-dev LexibombServer.Board

Maintains the state of game boards and provides the high-level public API for interacting with them throughout gameplay.

A default Lexibomb board consists of a 15 x 15 grid of squares. The squares are referred to using a coordinate system where the rows are numbered 1 through 15, and the columns are numbered “a” through “o”. A number combined with a letter specify a square on the board.

A “seed” value is stored with each board and used to seed the pseudorandom number generator (PRNG) whenever randomness is required. This adds the option for functional determinism when necessary.

Types

coord ::
  LexibombServer.Board.Grid.coord |
  {non_neg_integer, String.t} |
  String.t

seed :: {integer, integer, integer}

t :: %{grid: LexibombServer.Board.Grid.t, seed: seed}

Functions

get(pid)

Specs

get(pid) :: t

Retrieves the board’s state from the given process.

new(size \\ 15)

Specs

new(pos_integer) :: t

Creates a board of the given size.

A unique seed for the PRNG is generated and stored with the board.

new(size, seed)

Specs

new(pos_integer, seed) :: t

Creates a board of the given size with a specific PRNG seed.

parse_coord(coord)

Specs

parse_coord(coord) ::
  {:ok, LexibombServer.Board.Grid.coord} |
  {:error, :badarg}

Parses and normalizes a given coordinate.

It will parse any of the valid LexibombServer.Board.coord forms and will normalize them to the underlying LexibombServer.Board.Grid.coord type. If indicating the column with a string, you can use either uppercase or lowercase letters, and you may also optionally separate the row and column with whitespace.

Both row and col must parse into non-negative integers for success.

Examples

iex> LexibombServer.Board.parse_coord "10B"
{:ok, {10, 2}}
iex> LexibombServer.Board.parse_coord "10 b"
{:ok, {10, 2}}
iex> LexibombServer.Board.parse_coord {10, "B"}
{:ok, {10, 2}}
iex> LexibombServer.Board.parse_coord {10, 2}
{:ok, {10, 2}}
iex> LexibombServer.Board.parse_coord {-10, 2}
{:error, :badarg}
iex> LexibombServer.Board.parse_coord {"B", 10}
{:error, :badarg}

Returns {:ok, {row, col}} on success, or {:error, :badarg} on failure.

place_bomb(pid, coord)

Specs

place_bomb(pid, coord) :: :ok | {:error, :badarg}

Places a bomb on the board square at the given coordinate.

Returns :ok on success, or {:error, :badarg} on failure.

place_bombs(pid, coords)

Specs

place_bombs(pid, [coord]) :: :ok | {:error, :badarg}

Places a bomb on each of the board squares at the given coordinates.

Returns :ok on success, or {:error, :badarg} on failure.

place_random_bombs(pid, count \\ 22)

Specs

place_random_bombs(pid, pos_integer) :: :ok

Places a bomb on count number of randomly selected board squares.

The board’s PRNG seed is used to make this operation idempotent.

set(board, pid)

Specs

set(t, pid) :: :ok

Stores the state of board in the given process.

size(pid)

Specs

size(pid) :: pos_integer

Returns the size of the game board.

The “size”, is the number of squares along a single dimension of the board.

start_link()

Specs

start_link :: Agent.on_start

Starts an agent linked to the current process, storing the state of a new board.

Examples

iex> {:ok, pid} = LexibombServer.Board.start_link
iex> is_pid(pid)
true

start_link(board)

Specs

start_link(t) :: Agent.on_start

Starts an agent linked to the current process, storing the state of the given board.

valid_square?(pid, coord)

Specs

valid_square?(pid, coord) :: boolean

Validates if the given coordinate points to a square on the board.