Lexibomb Server v0.0.1-dev LexibombServer.Board

Maintains the state of game boards and provides the high-level public API for interacting with them throughout gameplay.

A default Lexibomb board consists of a 15 x 15 grid of squares. The squares are referred to using a coordinate system where the rows are numbered 1 through 15, and the columns are numbered “a” through “o”. A number combined with a letter specify a square on the board.

A “seed” value is stored with each board and used to seed the pseudorandom number generator (PRNG) whenever randomness is required. This adds the option for functional determinism when necessary.

Types

board_options :: [size: pos_integer, seed: seed, bomb_count: pos_integer]

coord ::
  LexibombServer.Board.Grid.coord |
  {non_neg_integer, String.t} |
  String.t

seed :: {integer, integer, integer}

t :: %{grid: LexibombServer.Board.Grid.t, seed: seed}

Functions

get(pid)

Specs

get(pid) :: t

Retrieves the board’s state from the given process.

make_play(pid, play)

Specs

make_play(pid, LexibombServer.Play.t) ::
  :ok |
  {:error, :badarg}

Make a play of tiles on the board.

new(opts \\ [])

Specs

new(board_options) :: t

Creates a new game board with the given options.

Options

The accepted options are:

:size - create a board of the given size; the default is 15

:seed - a seed for the PRNG; the default is a uniquely generated seed

:bomb_count - the number of bombs to randomly place on the board; the default is 22

parse_coord(coord)

Specs

parse_coord(coord) ::
  {:ok, LexibombServer.Board.Grid.coord} |
  {:error, :badarg}

Parses and normalizes a given coordinate.

It will parse any of the valid LexibombServer.Board.coord forms and will normalize them to the underlying LexibombServer.Board.Grid.coord type. If indicating the column with a string, you can use either uppercase or lowercase letters, and you may also optionally separate the row and column with whitespace.

Both row and col must parse into non-negative integers for success.

Examples

iex> LexibombServer.Board.parse_coord "10B"
{:ok, {10, 2}}
iex> LexibombServer.Board.parse_coord "10 b"
{:ok, {10, 2}}
iex> LexibombServer.Board.parse_coord {10, "B"}
{:ok, {10, 2}}
iex> LexibombServer.Board.parse_coord {10, 2}
{:ok, {10, 2}}
iex> LexibombServer.Board.parse_coord {-10, 2}
{:error, :badarg}
iex> LexibombServer.Board.parse_coord {"B", 10}
{:error, :badarg}

Returns {:ok, {row, col}} on success, or {:error, :badarg} on failure.

place_bomb(pid, coord)

Specs

place_bomb(pid, coord) :: :ok | {:error, :badarg}

Places a bomb on the board square at the given coordinate.

Returns :ok on success, or {:error, :badarg} on failure.

place_bombs(pid, coords)

Specs

place_bombs(pid, [coord]) :: :ok | {:error, :badarg}

Places a bomb on each of the board squares at the given coordinates.

Returns :ok on success, or {:error, :badarg} on failure.

place_random_bombs(board, count \\ 22)

Specs

place_random_bombs(t, pos_integer) :: t

Places a bomb on count number of randomly selected board squares.

The board’s PRNG seed is used to make this operation idempotent.

place_tile(pid, coord, tile)

Specs

place_tile(pid, coord, String.t) ::
  :ok |
  {:error, :badarg}

Places a tile on the board square at the given coordinate.

Returns :ok on success, or {:error, :badarg} on failure.

set(board, pid)

Specs

set(t, pid) :: :ok

Stores the state of board in the given process.

size(pid)

Specs

size(pid) :: pos_integer

Returns the size of the game board.

The “size”, is the number of squares along a single dimension of the board.

start_link(opts \\ [])

Starts an agent linked to the current process, storing the state of a new board.

Examples

iex> {:ok, pid} = LexibombServer.Board.start_link
iex> is_pid(pid)
true

Options

Same as new/1.

valid_square?(pid, coord)

Specs

valid_square?(pid, coord) :: boolean

Returns true if the given coordinate points to a square on the board.