Lexibomb Server v0.0.1-dev LexibombServer.Rack

Provides functions related to dealing with a rack of tiles.

A rack is a list of (usually 7) tiles, like ["E", "E", "L", "R", "T", "T", "S"].

Types

t :: %{tiles: [String.t]}

Functions

letters(rack)

Specs

letters(t) :: [String.t]

Returns the distinct letters in a rack.

Includes the lowercase alphabet if there are any blank tiles.

Examples

iex> LexibombServer.Rack.new("EELRTTS") |> LexibombServer.Rack.letters
["E", "L", "R", "T", "S"]
iex> LexibombServer.Rack.new("EELRTT_") |> LexibombServer.Rack.letters
["E", "L", "R", "T", "a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l",
 "m", "n", "o", "p", "q", "r", "s", "t", "u", "v", "w", "x", "y", "z"]

new(tiles)

Specs

new(String.t | [String.t]) :: t

Creates a new rack with the given tiles.

The tiles are normalized to uppercase.

tiles can be either a string, which will be tokenized for you, or a list of tiles.

Examples

iex> LexibombServer.Rack.new "Hello"
%LexibombServer.Rack{tiles: ["H", "E", "L", "L", "O"]}
iex> LexibombServer.Rack.new ~W(E E L R T T _)
%LexibombServer.Rack{tiles: ["E", "E", "L", "R", "T", "T", "_"]}

remove(rack, letters)

Specs

remove(t, String.t | [String.t]) :: t

Returns a copy of rack with the given letters removed.

If any of the given letters are lowercase, a corresponding blank tile is removed from the rack.

letters can be either a string, which will be tokenized for you, or a list of letters.

Examples

iex> LexibombServer.Rack.new("EELRTTS") |> LexibombServer.Rack.remove("SET")
%LexibombServer.Rack{tiles: ["E", "L", "R", "T"]}
iex> LexibombServer.Rack.new("EELRTT_") |> LexibombServer.Rack.remove(~W(T R E a T))
%LexibombServer.Rack{tiles: ["E", "L"]}