Lexibomb Server v0.0.1-dev LexibombServer.WordList

Manages access to the list of valid gameplay words.

Functions

member?(word)

Specs

member?(String.t) :: boolean

Checks if the word list contains word.

start_link(word_list \\ default_list)

Specs

start_link(MapSet.t) :: Agent.on_start

Starts an agent linked to the current process to store a normalized version of word_list.

The default word list is based on Mendel Leo Cooper’s Yet Another Word List (YAWL) project.